requirements.txt
python-telegram-bot==21.*
httpx==0.27.*
pydantic-settings==2.*
python-dotenv==1.0.*
